**Working on single threading
Create a list to hold an arbitrary number of values. Create a function named 'loop' to iterate through this list. Utilize and modify previous code that was creating and ending only one thread, to create x amount of threads according to the for loop. ('loop' function will be renamed to something more appropriate at a later stage.)